+++ apps/web/src/app/[locale]/(space)/account/profile/delete-account-dialog.tsx
@@ -15,8 +15,8 @@ import { Input } from "@rallly/ui/input";
 import { useForm } from "react-hook-form";
 
 import { Trans } from "@/components/trans";
-import { useSafeAction } from "@/features/safe-action/client";
 import { useTranslation } from "@/i18n/client";
+import { useSafeAction } from "@/lib/safe-action/client";
 import { deleteCurrentUserAction } from "./actions";
 
 export function DeleteAccountDialog({
