+++ apps/web/src/app/[locale]/admin-setup/make-me-admin-button.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from "@rallly/ui/button";
 import { Trans } from "@/components/trans";
-import { useSafeAction } from "@/features/safe-action/client";
+import { useSafeAction } from "@/lib/safe-action/client";
 import { makeMeAdminAction } from "./actions";
 
 export function MakeMeAdminButton() {
