+++ apps/web/src/app/[locale]/control-panel/settings/actions.ts
@@ -4,7 +4,7 @@ import { prisma } from "@rallly/database";
 import { revalidateTag } from "next/cache";
 import { instanceSettingsTag } from "@/features/instance-settings/constants";
 import { instanceSettingsSchema } from "@/features/instance-settings/schema";
-import { adminActionClient } from "@/features/safe-action/server";
+import { adminActionClient } from "@/lib/safe-action/server";
 
 export const updateInstanceSettingsAction = adminActionClient
   .metadata({
