+++ apps/web/src/features/spaces/components/space-dropdown.tsx
@@ -16,8 +16,8 @@ import { CirclePlusIcon, SettingsIcon } from "lucide-react";
 import Link from "next/link";
 import { ProBadge } from "@/components/pro-badge";
 import { Trans } from "@/components/trans";
-import { useSafeAction } from "@/features/safe-action/client";
 import { setActiveSpaceAction } from "@/features/spaces/actions";
+import { useSafeAction } from "@/lib/safe-action/client";
 import { NewSpaceDialog } from "./new-space-dialog";
 import { SpaceIcon } from "./space-icon";
